# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sever — peer-to-peer lending API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile/me:
    get:
      tags: [profile]
      operationId: getMyProfile
      summary: Get my profile, wallet balance, and trust score
      responses:
        '200':
          description: Profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyProfile'
    patch:
      tags: [profile]
      operationId: updateMyProfile
      summary: Update my display name and bio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileBody'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyProfile'
...
  /profile/withdraw:
    post:
      tags: [profile]
      operationId: withdrawFunds
      summary: Withdraw funds from wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawBody'
      responses:
        '200':
          description: New balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyProfile'
...
  /profile/username:
    put:
      tags: [profile]
      operationId: changeUsername
      summary: Set or update the user's unique username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUsernameBody'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyProfile'
        '409':
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /loans/{id}/messages:
    get:
      tags: [messages]
      operationId: listLoanMessages
      summary: Get the chat thread for a loan project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanMessage'
    post:
      tags: [messages]
      operationId: postLoanMessage
      summary: Post a message to a loan project chat
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostMessageBody'
      responses:
        '201':
          description: Created message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanMessage'
...
  /messages/{userId}:
    get:
      tags: [messages]
      operationId: getConversation
      summary: Get the DM thread with a specific user
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DirectMessage'
    post:
      tags: [messages]
      operationId: sendDirectMessage
      summary: Send a direct message to a user
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostMessageBody'
      responses:
        '201':
          description: Sent message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectMessage'
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
    MyProfile:
      type: object
      required:
        - id
        - username
        - displayName
        - bio
        - profileImageUrl
        - walletBalance
        - trustScore
        - tier
        - loansFunded
        - loansBorrowed
        - totalLent
        - totalBorrowed
        - onTimePayments
        - latePayments
      properties:
        id: { type: string }
        username: { type: [string, 'null'] }
        displayName: { type: string }
        bio: { type: [string, 'null'] }
        profileImageUrl: { type: [string, 'null'] }
        walletBalance: { type: number, description: "Available balance in USD" }
        trustScore: { type: integer, description: "0-1000 community trust score" }
        tier: { type: string, enum: [unverified, bronze, silver, gold, platinum] }
        loansFunded: { type: integer }
        loansBorrowed: { type: integer }
        totalLent: { type: number }
        totalBorrowed: { type: number }
        onTimePayments: { type: integer }
        latePayments: { type: integer }
    UpdateProfileBody:
      type: object
      required: [displayName]
      properties:
        displayName: { type: string, minLength: 1, maxLength: 60 }
        bio: { type: string, maxLength: 280 }
...
    WithdrawBody:
      type: object
      required: [amount]
      properties:
        amount: { type: number, minimum: 1 }
...
    SetUsernameBody:
      type: object
      required: [username]
      properties:
        username:
          type: string
    PostMessageBody:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1, maxLength: 1000 }
    LoanMessage:
      type: object
      required: [id, loanId, senderId, senderUsername, content, createdAt]
      properties:
        id: { type: string }
        loanId: { type: string }
        senderId: { type: string }
        senderUsername: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
    DirectMessage:
      type: object
      required: [id, fromUserId, toUserId, fromUsername, toUsername, content, read, createdAt]
      properties:
        id: { type: string }
        fromUserId: { type: string }
        toUserId: { type: string }
        fromUsername: { type: string }
        toUsername: { type: string }
        content: { type: string }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
...