# lib/api-zod/src/generated/types/directMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sever — peer-to-peer lending API
 * OpenAPI spec version: 0.1.0
 */

export interface DirectMessage {
  id: string;
  fromUserId: string;
  toUserId: string;
  fromUsername: string;
  toUsername: string;
  content: string;
  read: boolean;
  createdAt: Date;
}