# lib/api-zod/src/generated/types/loanMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sever — peer-to-peer lending API
 * OpenAPI spec version: 0.1.0
 */

export interface LoanMessage {
  id: string;
  loanId: string;
  senderId: string;
  senderUsername: string;
  content: string;
  createdAt: Date;
}