# lib/api-zod/src/generated/types/myProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sever — peer-to-peer lending API
 * OpenAPI spec version: 0.1.0
 */
import type { MyProfileTier } from "./myProfileTier";

export interface MyProfile {
  id: string;
  /** @nullable */
  username: string | null;
  displayName: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  /** Available balance in USD */
  walletBalance: number;
  /** 0-1000 community trust score */
  trustScore: number;
  tier: MyProfileTier;
  loansFunded: number;
  loansBorrowed: number;
  totalLent: number;
  totalBorrowed: number;
  onTimePayments: number;
  latePayments: number;
}