# lib/api-zod/src/generated/types/postMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sever — peer-to-peer lending API
 * OpenAPI spec version: 0.1.0
 */

export interface PostMessageBody {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  content: string;
}